+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.0.5
+- Fix: arrange tokens on ladders
+- AI: a bit more intelligent, checks if a token can move on a bench or kill
+
 ## v0.0.4
 - Fix: some refactoring
 - Feature: Show triple 6 message
@@ -5,7 +9,6 @@
 - Add fonts
 
 ## v0.0.3
-
 - Feature: Throw again with a 6
 - Feature: Send all home with 3x6
 - Fix: Fix mismatch horizontal ladders